Fix slider FHIR export: integer type + questionnaire-itemControl extension
- slider type now maps to FHIR integer (not decimal) — pain/VAS scales are
  integer-valued; minValue/maxValue extensions use valueInteger accordingly
- Adds questionnaire-itemControl: slider extension per FHIR SDC spec so any
  SDC-compliant delivery tool renders a slider widget; LHC-Forms falls back
  to a validated integer text input (slider itemControl unsupported there)

+++ tests/test_fhir_export.py
@@ -153,7 +153,7 @@ def test_empty_questionnaire_has_no_item_key(tmp_path):
     ("date",            "date"),
     ("datetime",        "dateTime"),
     ("likert",          "choice"),
-    ("slider",          "decimal"),
+    ("slider",          "integer"),
 ])
 def test_type_mapping(tmp_path, qtype, expected_fhir):
     conn = _db(tmp_path)
